Descarta dropdowns da topbar
