+++ .env
@@ -1,1 +1,1 @@
-VITE_API_URL=http://localhost:3000
+VITE_API_URL=http://localhost:3001
